+++ R/plotCrossHeatmap.R
@@ -32,7 +32,21 @@ extractCrossInferenceData <- function(resLs){
       table <- summary(mdl)$s.table %>% as.data.frame()
       table$coefficient <- rownames(table)
       table$combination <- x
-      table <- table %>% separate(.data[["combination"]], c("cell1", "cell2"), sep = "_")
+      table <- table %>% separate(.data[["combination"]],
+                                  c("cell1", "cell2"), sep = "_")
+
+      #extract the mean functional coefficient as effect size measure
+      coef <- coef(mdl)
+      df <- lapply(rownames(table), function(predictor){
+        df <- coef$sm[[predictor]]$coef
+        df[["coefficient"]] <- predictor
+        df <- df %>% mutate(mean_coefficient = mean(.data[["value"]])) %>%
+          select(.data[["mean_coefficient"]], .data[["coefficient"]]) %>%
+          unique()
+        return(df)
+      }) %>% dplyr::bind_rows()
+      #join by coefficient
+      table <- table %>% dplyr::left_join(df, by = "coefficient")
       return(table)
     }
   }) %>% dplyr::bind_rows()
@@ -81,18 +95,23 @@ plotCrossHeatmap <- function(resLs,
    df <- df %>% filter(.data[["coefficient"]] %in% coefficientsToPlot)
  }
  if(is.null(adj.pvalue)){
-   p <- ggplot(df, aes(.data[["cell1"]], .data[["cell2"]],
-                       fill = -log10(.data[["p-value"]] + 0.001))) +
-     geom_tile(colour="white", size=0.2) +
-     facet_wrap(~.data[["coefficient"]]) +
-     theme_light()
+   p <- ggplot(df, aes(x = .data[["cell1"]], y = .data[["cell2"]])) +
+     geom_point(aes(color = .data[["mean_coefficient"]],
+                    size = -log10(.data[["p-value"]] + 0.001)))
  }else{
    df[["adj.p-value"]] <- stats::p.adjust(df[["p-value"]], method = adj.pvalue)
-   p <- ggplot(df, aes(.data[["cell1"]], .data[["cell2"]],
-                       fill = -log10(.data[["adj.p-value"]] + 0.001))) +
-     geom_tile(colour="white", size=0.2, ...) +
-     facet_wrap(~.data[["coefficient"]], ...) +
-     theme_light()
+   p <- ggplot(df, aes(x = .data[["cell1"]], y = .data[["cell2"]])) +
+     geom_point(aes(color = .data[["mean_coefficient"]],
+                    size = -log10(.data[["adj.p-value"]] + 0.001)))
  }
+ p <- p + scale_x_discrete(guide = guide_axis(angle = 50)) +
+   geom_point(aes(size = -log10(.data[["p-value"]] + 0.001)),
+              shape = 1,colour = "black")+
+   facet_wrap(~.data[["coefficient"]]) +
+   theme_light() +
+   scale_colour_gradient2(midpoint = 0,
+                          high = scales::muted("red"),
+                          mid = "white",
+                          low = scales::muted("blue"))
  return(p)
 }